pyproject.toml
[project]
name = "quad"
version = "1.2.0"
description = "Quake Ultimate Analysis and Development. A system to record, analyze, organize Quake Champions recordings and improve in Quake in general"
dependencies = [
	"cffi",
	"flask",
	"google-api-python-client",
	"oauth2client",
	"opencv-python",
	"pexpect",
	"py-cord",
	"pytesseract",
	"zmq",
	"Flask-SQLAlchemy",
	"Flask-Migrate",
	"coverage",
	"pytest",
	"pytest-mock"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = false
log_cli_level = "INFO"

[tool.coverage.run]
branch = true
source = ["quad"]
omit = [
	"quad/ndi/*"
]
